added authentication fror update and delete of customer

+++ Backend/src/controllers/customerController.js
@@ -29,7 +29,7 @@ export const createCustomer = async (req, res) => {
 
 export const getCustomers = async (req, res) => {
     try {
-        
+
         const customers = await CustomerModel.find({ createdBy: req.user._id }).select('-__v');// Only fetch customers created by this user
         res.json({ customers });
 
@@ -47,7 +47,7 @@ export const getCustomers = async (req, res) => {
 
 export const getACustomer = async (req, res) => {
     try {
-        const customer = await CustomerModel.findById({_id: req.params.id, createdBy: req.user._id});
+        const customer = await CustomerModel.findById({ _id: req.params.id, createdBy: req.user._id });
         res.status(200).json({ customer })
     } catch (error) {
         console.log(error.message);
@@ -63,7 +63,7 @@ export const getACustomer = async (req, res) => {
 export const updateCustomer = async (req, res) => {
     try {
 
-        const customer = await CustomerModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        const customer = await CustomerModel.findByIdAndUpdate({ _id: req.params.id, createdBy: req.user._id }, req.body, { new: true });//ensure the task belong to the loggedin user
 
         res.json({ updated: customer });
 
@@ -80,12 +80,12 @@ export const updateCustomer = async (req, res) => {
 
 export const deleteCustomer = async (req, res) => {
     try {
-        await CustomerModel.findByIdAndDelete(req.params.id);
+        await CustomerModel.findByIdAndDelete({ _id: req.params.id, createdBy: req.user._id });//ensure the customer belong to the loggedin user
         res.json({ message: "Customer deleted" });
     } catch (error) {
         console.log(error.message);
         res.status(500).json({
-            message: "Something went wrong while adding data",
+            message: "Something went wrong while deleting data",
             error: error.message
         })
     }
